requirements.txt
streamlit==1.31.0
assemblyai
openai
python-dotenv


